Lavet vægt til double

+++ src/Models/Food.java
@@ -6,7 +6,7 @@ public class Food extends Item {
     public int healthPoints;
 
     public Food(String name, String DESCRIPTION, int healthPoints){
-        super(name, DESCRIPTION, new Random().nextInt(1,10));
+        super(name, DESCRIPTION, new Random().nextDouble(1,10));
         this.healthPoints = healthPoints;
     }
 }
