+++ src/Models/Weapon.java
@@ -5,7 +5,7 @@ import java.util.Random;
 public class Weapon extends Item {
 
     public Weapon(String name, String DESCRIPTION){
-        super(name, DESCRIPTION, new Random().nextInt(10,20));
+        super(name, DESCRIPTION, new Random().nextDouble(10,20));
     }
 
 }
